ingredients main functional

+++ db/migrate/20180414170815_ingredients_add_protein.rb
@@ -1,8 +1,8 @@
 class IngredientsAddProtein < ActiveRecord::Migration[5.1]
   def change
-    add_column :ingredients, :proteins, :integer
-    add_column :ingredients, :fats, :integer
-    add_column :ingredients, :carbohydrates, :integer
-    add_column :ingredients, :calories, :integer
+    add_column :ingredients, :proteins, :float
+    add_column :ingredients, :fats, :float
+    add_column :ingredients, :carbohydrates, :float
+    add_column :ingredients, :calories, :float
   end
 end
